Reland "Upconvert various types to int.", neteq portion.
This reverts portions of commit cb180976dd0e9672cde4523d87b5f4857478b5e9, which
reverted commit 83ad33a8aed1fb00e422b6abd33c3e8942821c24.  Specifically, the
files in webrtc/modules/audio_coding/neteq/ are relanded.

The original commit message is below:

Upconvert various types to int.

Per comments from HL/kwiberg on https://webrtc-codereview.appspot.com/42569004 , when there is existing usage of mixed types (int16_t, int, etc.), we'd prefer to standardize on larger types like int and phase out use of int16_t.

Specifically, "Using int16 just because we're sure all reasonable values will fit in 16 bits isn't usually meaningful in C."

This converts some existing uses of int16_t (and, in a few cases, other types such as uint16_t) to int (or, in a few places, int32_t).  Other locations will be converted to size_t in a separate change.

BUG=none
TBR=kwiberg

Review URL: https://codereview.webrtc.org/1181073002

Cr-Commit-Position: refs/heads/master@{#9427}

+++ webrtc/modules/audio_coding/neteq/expand.h
@@ -72,7 +72,7 @@ class Expand {
 
   void GenerateBackgroundNoise(int16_t* random_vector,
                                size_t channel,
-                               int16_t mute_slope,
+                               int mute_slope,
                                bool too_many_expands,
                                size_t num_noise_samples,
                                int16_t* buffer);
@@ -113,7 +113,7 @@ class Expand {
     AudioVector expand_vector0;
     AudioVector expand_vector1;
     bool onset;
-    int16_t mute_slope; /* Q20 */
+    int mute_slope; /* Q20 */
   };
 
   // Calculate the auto-correlation of |input|, with length |input_length|
@@ -123,7 +123,7 @@ class Expand {
   void Correlation(const int16_t* input,
                    size_t input_length,
                    int16_t* output,
-                   int16_t* output_scale) const;
+                   int* output_scale) const;
 
   void UpdateLagIndex();
